Separate access check functions into a separate module
