+++ crates/postgres-subsystem/postgres-resolver/src/aggregate_query.rs
@@ -8,8 +8,8 @@
 // by the Apache License, Version 2.0.
 
 use super::{
-    postgres_execution_error::PostgresExecutionError, sql_mapper::SQLOperationKind,
-    util::check_access,
+    auth_util::check_access, postgres_execution_error::PostgresExecutionError,
+    sql_mapper::SQLOperationKind,
 };
 use crate::operation_resolver::OperationSelectionResolver;
 use async_recursion::async_recursion;
