+++ crates/postgres-subsystem/postgres-resolver/src/postgres_mutation.rs
@@ -8,9 +8,10 @@
 // by the Apache License, Version 2.0.
 
 use super::{
+    auth_util::check_access,
     postgres_execution_error::PostgresExecutionError,
     sql_mapper::SQLOperationKind,
-    util::{check_access, find_arg, return_type_info},
+    util::{find_arg, return_type_info},
 };
 use crate::{
     create_data_param_mapper::InsertOperation, operation_resolver::OperationResolver,
